+++ jest.config.cjs
@@ -7,10 +7,13 @@ module.exports = {
     setupFiles: [
         '<rootDir>/test/bootstrap.cjs'
     ],
-    testURL: 'http://localhost:8080',
+    testEnvironmentOptions: {
+        url: 'http://localhost:8080',
+    },
     moduleNameMapper: {
         '^(.*)\.js$': '$1',
     },
+    transform: {
+        '\\.[jt]sx?$': ['babel-jest', { configFile: './test/babel.config.cjs' }]
+    },
 };
-
-